Split into three tools
